+++ src/commands.cpp
@@ -12,7 +12,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
- * You should have received a std::copy of the GNU General Public License
+ * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
@@ -25,6 +25,7 @@
 #include <map>
 #include <set>
 #include <string>
+#include <utility>
 #include <vector>
 
 #ifndef __cpp_lib_make_unique
@@ -508,29 +509,6 @@ Commands::Match Commands::find_command(Shell &shell, const CommandLine &command_
 	}
 
 	return commands;
-}
-
-void Commands::for_each_available_command(Shell &shell, apply_function f) const {
-	auto commands = commands_.equal_range(shell.context());
-
-	for (auto command_it = commands.first; command_it != commands.second; command_it++) {
-		if (shell.has_flags(command_it->second.flags_, command_it->second.not_flags_)) {
-			std::vector<std::string> name;
-			std::vector<std::string> arguments;
-
-			name.reserve(command_it->second.name_.size());
-			for (auto flash_name : command_it->second.name_) {
-				name.push_back(std::move(read_flash_string(flash_name)));
-			}
-
-			arguments.reserve(command_it->second.arguments_.size());
-			for (auto flash_argument : command_it->second.arguments_) {
-				arguments.push_back(std::move(read_flash_string(flash_argument)));
-			}
-
-			f(name, arguments);
-		}
-	}
 }
 
 Commands::Command::Command(unsigned int flags, unsigned int not_flags,